refactor(auth-context): move auth context types and mock to shared
