+++ apps/frontend/src/components/client/user/ProfileSection.tsx
@@ -2,6 +2,7 @@
 
 import { isGender, isPlayLevel } from "@repo/shared"
 import type { User } from "@repo/shared/payload-types"
+import type { AuthContextValue } from "@repo/shared/types/auth"
 import {
   AdditionalInfo,
   AdditionalInfoFields,
@@ -17,7 +18,6 @@ import {
 import { ConfirmationPopUp } from "@repo/ui/components/Generic"
 import { Container, Grid, GridItem, useDisclosure, useNotice } from "@yamada-ui/react"
 import { memo, useState } from "react"
-import type { AuthContextValue } from "@/context/AuthContext"
 import { useDeleteBooking } from "@/services/admin/bookings/AdminBookingMutations"
 import { useUpdateSelfMutation } from "@/services/auth/useUpdateSelfMutation"
 import { useMyBookings } from "@/services/bookings/BookingQuery"
