+++ packages/shared/src/types/auth/index.ts
@@ -1,2 +1,3 @@
+export * from "./context"
 export * from "./login"
 export * from "./register"
